chore: relocate tetst to the PicData modules
