+++ src/pic/data/grayscale.rs
@@ -92,3 +92,23 @@ impl PicData for GrayscaleData {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_pic_new_grayscale() {
+        let mut rng = StdRng::from_rng(rand::thread_rng()).unwrap();
+        let pic = GrayscaleData::new(0, 60, false, &mut rng, &vec![&"eye.jpg".to_string()]);
+        match &pic {
+            Pic::Grayscale(GrayscaleData { c, coord: _coord }) => {
+                let len = c.get_children().unwrap().len();
+                assert!(len > 0 && len < 60);
+            }
+            _ => {
+                panic!("wrong type");
+            }
+        };
+    }
+}
