+++ src/pic/pic.rs
@@ -416,103 +416,8 @@ mod tests {
     use super::*;
     use crate::parser::token::Token;
     use crate::pic::coordinatesystem::DEFAULT_COORDINATE_SYSTEM;
-    use image::buffer::Pixels;
     use image::io::Reader as ImageReader;
-    use image::{DynamicImage, GenericImageView, ImageBuffer, ImageFormat};
-    use std::io::{Cursor, Read};
-
-    #[test]
-    fn test_pic_new_mono() {
-        let mut rng = StdRng::from_rng(rand::thread_rng()).unwrap();
-        let pic = MonoData::new(0, 60, false, &mut rng, &vec![&"eye.jpg".to_string()]);
-        match &pic {
-            Pic::Mono(MonoData { c, coord: _coord }) => {
-                let len = c.get_children().unwrap().len();
-                assert!(len > 0 && len < 60);
-            }
-            _ => {
-                panic!("wrong type");
-            }
-        };
-    }
-
-    #[test]
-    fn test_pic_new_grayscale() {
-        let mut rng = StdRng::from_rng(rand::thread_rng()).unwrap();
-        let pic = GrayscaleData::new(0, 60, false, &mut rng, &vec![&"eye.jpg".to_string()]);
-        match &pic {
-            Pic::Grayscale(GrayscaleData { c, coord: _coord }) => {
-                let len = c.get_children().unwrap().len();
-                assert!(len > 0 && len < 60);
-            }
-            _ => {
-                panic!("wrong type");
-            }
-        };
-    }
-
-    #[test]
-    fn test_pic_new_gradient() {
-        let mut rng = StdRng::from_rng(rand::thread_rng()).unwrap();
-        let pic = GradientData::new(0, 60, false, &mut rng, &vec![&"eye.jpg".to_string()]);
-        match &pic {
-            Pic::Gradient(GradientData {
-                colors,
-                index,
-                coord: _coord,
-            }) => {
-                let len = colors.len();
-                assert!(len > 1 && len < 10);
-                let len = index.get_children().unwrap().len();
-                assert!(len > 0 && len < 60);
-            }
-            _ => {
-                panic!("wrong type");
-            }
-        };
-    }
-
-    #[test]
-    fn test_pic_new_rgb() {
-        let mut rng = StdRng::from_rng(rand::thread_rng()).unwrap();
-        let pic = RGBData::new(0, 60, false, &mut rng, &vec![&"eye.jpg".to_string()]);
-        match &pic {
-            Pic::RGB(RGBData { r, g, b, coord: _ }) => {
-                let len = r.get_children().unwrap().len();
-                assert!(len > 0 && len < 60);
-
-                let len = g.get_children().unwrap().len();
-                assert!(len > 0 && len < 60);
-
-                let len = b.get_children().unwrap().len();
-                assert!(len > 0 && len < 60);
-            }
-            _ => {
-                panic!("wrong type");
-            }
-        };
-    }
-
-    #[test]
-    fn test_pic_new_hsv() {
-        let mut rng = StdRng::from_rng(rand::thread_rng()).unwrap();
-        let pic = HSVData::new(0, 60, false, &mut rng, &vec![&"eye.jpg".to_string()]);
-        match &pic {
-            Pic::HSV(HSVData { h, s, v, coord: _ }) => {
-                let len = h.get_children().unwrap().len();
-                assert!(len > 0 && len < 60);
-
-                let len = s.get_children().unwrap().len();
-                assert!(len > 0 && len < 60);
-
-                let len = v.get_children().unwrap().len();
-                assert!(len > 0 && len < 60);
-            }
-            _ => {
-                panic!("wrong type");
-            }
-        };
-    }
+    use image::{DynamicImage, GenericImageView, ImageBuffer};
 
     #[test]
     fn test_pic_to_lisp() {
